% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{annotate_regression}
\alias{annotate_regression}
\title{annotate_regression}
\usage{
annotate_regression(lmfit, ggp, font_size = 10)
}
\arguments{
\item{lmfit}{an linear model fit object}

\item{ggp}{a ggplot2 object}
}
\description{
This function parses an lmfit object, extracts the partial
correlation between the last predictor of the model and the outcome,
and adds it to a ggplot object in the right hand corner. Assumes
that the variables are standardized and that the predictor of
interest is the last one in the model.
}
\keyword{annotation}
\keyword{ggplot2,}
\keyword{regression,}
