% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{create_package}
\alias{create_package}
\title{create_package}
\usage{
create_package(project_name)
}
\arguments{
\item{project_name}{Name of the project.}
}
\description{
This function creates an R package from scratch.
}
\examples{
create_package("myproject", "~/tmp")
}
\keyword{R}
\keyword{module}
