% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{create_project}
\alias{create_project}
\title{create_project}
\usage{
create_project(tags = list(name = "myproject", title = "My Project",
  author = "Firstname Lastname", email = "lastname at inst dot edu", inst =
  "My Institution", root = "~\\\\/projects"))
}
\arguments{
\item{name}{The name of the project}

\item{title}{The title of the project}

\item{author}{The author of the project}

\item{email}{The email of the project's author}

\item{inst}{The institution of the project's author}

\item{root}{The root folder for the project}
}
\description{
This function sets up the typical file structure for projects
}
\examples{
create_project("myproject", "My Author", "myemail@me.com",
    "My Institution", "rootdirectory")
}
\keyword{R}
\keyword{module}
\keyword{reproducible}
\keyword{research}
