% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{dic}
\alias{dic}
\title{dic}
\usage{
dic(stanfit)
}
\arguments{
\item{stanfit}{Stan object.}
}
\description{
This function calculates the Deviance Information Criterion (DIC).
}
\examples{
dic(stanfit)
}
\keyword{stan}
