% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{get_partcorr_vec}
\alias{get_partcorr_vec}
\title{get_partcorr_vec}
\usage{
get_partcorr_vec(lmfit, xc)
}
\arguments{
\item{lmfit}{Linear model object}

\item{xc}{Name of predictor x}
}
\description{
This function returns partial correlation vectors
}
\keyword{correlation}
\keyword{linear}
\keyword{model,}
\keyword{partial}
