% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_chi}
\alias{parse_chi}
\title{parse_chi}
\usage{
parse_chi(chisq, def)
}
\arguments{
\item{chisq}{Test result}

\item{def}{Degrees of freedom}
}
\description{
This function parses and formats a chi square test result
}
\keyword{anova}
\keyword{stats,}
