% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_currency}
\alias{parse_currency}
\title{parse_currency}
\usage{
parse_currency(amount, currency)
}
\arguments{
\item{amount}{amount}

\item{currency}{currency}
}
\description{
This function will parse a currency amount
}
\keyword{org-mode,}
\keyword{table}
