% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_effectsize}
\alias{parse_effectsize}
\title{parse_effectsize}
\usage{
parse_effectsize(es, type = "d", digits = 2)
}
\arguments{
\item{es}{d effect size}

\item{type}{type of effect size}

\item{digits}{number of decimal places}
}
\description{
This function parses and formats an effect size
}
\keyword{stats}
