% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_estci}
\alias{parse_estci}
\title{parse_estci}
\usage{
parse_estci(lsm)
}
\arguments{
\item{dat}{Object from lsmeans.}
}
\description{
This function parses and formats an lsmeans estimate with 95% CI.
}
\keyword{confidence}
\keyword{estimate,}
\keyword{interval}
\keyword{stats,}
