% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_fstat}
\alias{parse_fstat}
\title{parse_fstat}
\usage{
parse_fstat(anovatab, index)
}
\arguments{
\item{anovatab}{anova table object}

\item{index}{Line index in data frame}
}
\description{
This function parses and formats an anova table
}
\examples{
anovatab <- anova(lm(rnorm(100, 0, 1) ~ rnorm(100, 0, 1)))
parse_fstat(anovatab, 1)
}
\keyword{anova}
\keyword{stats,}
