% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_lm}
\alias{parse_lm}
\title{parse_lm}
\usage{
parse_lm(lmfit, index = 2, scaled = FALSE, style = "beta")
}
\arguments{
\item{lmfit}{lm object}

\item{index}{Line index in data frame}

\item{scaled}{Was the predicor scaled? Defaults to FALSE.}

\item{style}{Output format. Defaults to beta.}
}
\description{
This function parses and formats an linear model object.
}
\examples{
lmfit <- lm(rnorm(100, 0, 1) ~ rnorm(100, 0, 1))
parse_lm(lmfit, 1, FALSE, "beta")
}
\keyword{linear}
\keyword{model}
\keyword{stats,}
