% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_lmer}
\alias{parse_lmer}
\title{parse_lmer}
\usage{
parse_lmer(lmerfit, ind1, ind2, sqrt_b = FALSE)
}
\arguments{
\item{lmerfit}{lmer object.}

\item{ind1}{First index.}

\item{ind2}{Second index.}

\item{sqrt_b}{Take the square root of the coefficient. Default FALSE.}
}
\description{
This function parses and formats an lmerfit object
}
\keyword{lme4,}
\keyword{mixed}
\keyword{models}
\keyword{org-mode,}
\keyword{parsing,}
\keyword{text,}
