% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_msd}
\alias{parse_msd}
\title{parse_msd}
\usage{
parse_msd(m, sd)
}
\arguments{
\item{m}{Mean}

\item{sd}{Standard deviation}
}
\description{
This function parses and formats a mean and SD.
}
\examples{
parse_msd(0, 1)
}
\keyword{format}
