% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_propstat}
\alias{parse_propstat}
\title{parse_propstat}
\usage{
parse_propstat(propstat)
}
\arguments{
\item{propstat}{prop.test object}
}
\description{
This function parses and formats a equal proportions statistic.
}
\examples{
propstat <- prop.test(c(100, 110), c(200, 200))
parse_propstat(propstat)
}
\keyword{proportions}
\keyword{stats,}
