% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_pval}
\alias{parse_pval}
\title{parse_pval}
\usage{
parse_pval(pval, rf = 2, chs = "steq")
}
\arguments{
\item{pval}{P-value}

\item{rf}{Decimal point for rounding}

\item{chs}{String for equality test. Defaults to "steq" ("<=")}
}
\description{
This function parses and formats a P-value
}
\examples{
parse_pval(0.0001, 2)
}
\keyword{org-mode,}
\keyword{stats}
