% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_pwr}
\alias{parse_pwr}
\title{parse_pwr}
\usage{
parse_pwr(pwr, param = "d", digits = 2)
}
\arguments{
\item{pwr}{object from pwr library}

\item{param}{type of variable}

\item{digits}{number of decimal places}
}
\description{
This function parses and formats a pwr object
}
\keyword{stats}
