% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_rstat}
\alias{parse_rstat}
\title{parse_rstat}
\usage{
parse_rstat(rstat, method = "pearson")
}
\arguments{
\item{rstat}{cor.test object}

\item{method}{Type of correlation. Defaults to Pearson.}
}
\description{
This function parses and formats an r statistic.
}
\examples{
rstat <- cor.test(rnorm(100, 0, 1), rnorm(100, 0, 1))
parse_rstat(rstat)
}
\keyword{correlation}
\keyword{stats,}
