% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_table}
\alias{parse_table}
\title{parse_table}
\usage{
parse_table(tabdf)
}
\arguments{
\item{tabdf}{data frame}
}
\description{
This function removes any nils by replacing NA by whitespace
}
\keyword{org-mode,}
\keyword{table}
