% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_template}
\alias{parse_template}
\title{parse_template}
\usage{
parse_template(filename, tags)
}
\arguments{
\item{filename}{File to parse.}

\item{tags}{List of tags.}
}
\description{
This function parses a file and replaces all @tags@.
}
\examples{
parse_template("mytemplate.txt", list("author"="My Author")
}
\keyword{parsing,}
\keyword{template}
