% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_tstat}
\alias{parse_tstat}
\title{parse_tstat}
\usage{
parse_tstat(tstat)
}
\arguments{
\item{tstat}{t-test object}
}
\description{
This function parses and formats a t statistic.
}
\examples{
tstat <- t.test(rnorm(100, 0, 1))
parse_tstat(tstat)
}
\keyword{stats,}
\keyword{t-test}
