% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{parse_vals}
\alias{parse_vals}
\title{parse_vals}
\usage{
parse_vals(action = "pval", param)
}
\arguments{
\item{action}{action string. Defaults to pval}

\item{param}{P-value}
}
\description{
This function parses and formats a P-value.
}
\examples{
parse_vals("pval", 0.0001)
}
\keyword{P-value}
