% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{rand_str}
\alias{rand_str}
\title{rand_str}
\usage{
rand_str(n = 1)
}
\arguments{
\item{n}{Number of samples.}
}
\description{
This function returns a random string.
}
\examples{
return 1000 random strings
rand_fn(1000)
}
\keyword{parsing,}
\keyword{random}
\keyword{template,}
