% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{starsfromp}
\alias{starsfromp}
\title{starsfromp}
\usage{
starsfromp(pval, c1 = "~", c2 = "*")
}
\arguments{
\item{pval}{p-value of interest.}

\item{c1}{character for P < 0.1.}

\item{c2}{character(s) for P < 0.05.}
}
\description{
This function returns asterisks for your P-value.
}
\examples{
starsfromp(0.02, "")
}
\keyword{P-value}
