% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{create_project}
\alias{create_project}
\title{create_project}
\usage{
create_project(tags = list(name = "myproject", title = "My Project",
  shorttitle = "My short title", author = "Firstname Lastname", email =
  "lastname at inst dot edu", git =
  "https:\\\\/\\\\/github.com\\\\/mypath", bibliography = "master",
  bibstyle = "nature", keywords = "git; org-mode; R", date = date(), inst =
  "My Institution", root = "~\\\\/projects"), flavor = "org-mode")
}
\arguments{
\item{name}{The name of the project}

\item{title}{The title of the project}

\item{author}{The author of the project}

\item{email}{The email of the project's author}

\item{inst}{The institution of the project's author}

\item{root}{The root folder for the project}
}
\description{
This function sets up the typical file structure for projects
}
\examples{
create_project(tags=list("date"=date()), flavor="org-mode")
}
\keyword{R}
\keyword{module}
\keyword{reproducible}
\keyword{research}
