% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represearch.R
\name{represearch_slides}
\alias{represearch_slides}
\title{represearch_slides}
\usage{
represearch_slides()
}
\description{
This function dowloads the pdf presentation
}
\keyword{reproducible}
\keyword{research}
